feat: Change fix activity stack, open comment layout, password layout
