+++ app/src/main/java/com/doublejj/edit/ui/modules/main/myedit/settings/withdrawal/WithdrawalCompleteActivity.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.databinding.DataBindingUtil
+import com.doublejj.edit.ApplicationClass
 import com.doublejj.edit.R
 import com.doublejj.edit.databinding.ActivityWithdrawalCompleteBinding
 import com.doublejj.edit.ui.modules.main.splash.SplashActivity
@@ -17,6 +18,9 @@ class WithdrawalCompleteActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = DataBindingUtil.setContentView(this, R.layout.activity_withdrawal_complete)
+
+        // add activity at sActivityList
+        ApplicationClass.sActivityList.add(this)
 
         /** toolbar buttons **/
         binding.ibBack.setOnClickListener {
@@ -38,6 +42,23 @@ class WithdrawalCompleteActivity : AppCompatActivity() {
             sendIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
             startActivity(sendIntent)
         }
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+
+        val editor = ApplicationClass.sSharedPreferences.edit()
+        editor.putString(ApplicationClass.X_ACCESS_TOKEN, null)
+        editor.putString(ApplicationClass.USER_POSITION, null)
+        editor.putBoolean(ApplicationClass.MENTOR_AUTH_CONFIRM, false)
+        editor.putString(ApplicationClass.USER_NICKNAME, null)
+        editor.putString(ApplicationClass.USER_EMOTION, null)
+        editor.putString(ApplicationClass.USER_COLOR, null)
+        editor.commit()
+        editor.apply()
+
+        ApplicationClass.sActivityList.actFinish()
+
     }
 
 }
