+++ app/src/main/java/com/doublejj/edit/ui/modules/main/myedit/switch_position/SwitchPositionCompleteActivity.kt
@@ -19,6 +19,9 @@ class SwitchPositionCompleteActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         binding = DataBindingUtil.setContentView(this, R.layout.activity_switch_position_complete)
 
+        // add activity at sActivityList
+        ApplicationClass.sActivityList.add(this)
+
         /** toolbar buttons **/
         binding.ibBack.setOnClickListener {
             // 초기화면으로 가기
@@ -33,11 +36,12 @@ class SwitchPositionCompleteActivity : AppCompatActivity() {
         val spanStr = CustomSpannableString(applicationContext).getPurpleActiveColorText(textTitle, nickName, R.color.purple_active)
         binding.tvCompleteTitle.setText(spanStr)
         
-        binding.btnRelogin.setOnClickListener { 
+        binding.btnRelogin.setOnClickListener {
             // 초기화면으로 가기
             val sendIntent = Intent(this, SplashActivity::class.java)
             sendIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK)
             startActivity(sendIntent)
+
         }
     }
 
@@ -53,5 +57,8 @@ class SwitchPositionCompleteActivity : AppCompatActivity() {
         editor.putString(ApplicationClass.USER_COLOR, null)
         editor.commit()
         editor.apply()
+
+        ApplicationClass.sActivityList.actFinish()
+
     }
 }
